Module Ppx_js__Ppx_js_internal

val nolabel : Migrate_parsetree.OCaml_408.Ast.Asttypes.arg_label

val may_tuple : ?⁠loc:'a -> (?⁠loc:'a -> ?⁠attrs:'b -> 'c list -> 'c) -> 'c list -> 'c option

val lid : ?⁠loc:Migrate_parsetree.OCaml_408.Ast.Ast_helper.loc -> string -> Migrate_parsetree.OCaml_408.Ast.Longident.t Migrate_parsetree.OCaml_408.Ast.Location.loc

val constr : ?⁠loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> ?⁠attrs:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.attrs -> string -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val unit : ?⁠loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> ?⁠attrs:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.attrs -> unit -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val tuple : ?⁠loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> ?⁠attrs:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.attrs -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val str : ?⁠loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> ?⁠attrs:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.attrs -> string -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val exp_to_string : Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> string

Check if an expression is an identifier and returns it. Raise a Location.error if it's not.

val lid : ?⁠loc:Migrate_parsetree.OCaml_408.Ast.Ast_helper.loc -> string -> Migrate_parsetree.OCaml_408.Ast.Longident.t Migrate_parsetree.OCaml_408.Ast.Location.loc

val typ : string -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type

val arrows : (Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Asttypes.arg_label * Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type) list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type

arg1 -> arg2 -> ... -> ret

val wrapper : string option Stdlib.ref

val make_str : ?⁠loc:Migrate_parsetree.OCaml_408.Ast.Location.t -> 'a -> 'a Migrate_parsetree.OCaml_408.Ast.Location.loc

val inside_Js : bool lazy_t

val merlin_noloc : Migrate_parsetree.OCaml_408.Ast.Parsetree.attribute

module Js : sig ... end

val unescape : string -> string

val app_arg : 'a -> Migrate_parsetree.OCaml_408.Ast.Asttypes.arg_label * 'a

val inject_arg : Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> Migrate_parsetree.OCaml_408.Ast.Parsetree.expression

val inject_args : Migrate_parsetree.OCaml_408.Ast.Parsetree.expression list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

module Arg : sig ... end

val js_dot_t_the_first_arg : Arg.t list -> (Migrate_parsetree.OCaml_408.Ast.Asttypes.arg_label * Migrate_parsetree.OCaml_408.Ast.Parsetree.core_type) list

val invoker : ?⁠extra_types:Arg.t list -> (Arg.t list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type) -> (Arg.t list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type -> (Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Asttypes.arg_label * Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type) list * Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.core_type) -> (Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression) -> Arg.t list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val open_t : Migrate_parsetree.OCaml_408.Ast.Ast_helper.loc -> Migrate_parsetree.OCaml_408.Ast.Parsetree.core_type

val method_call : loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> string -> (Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Asttypes.arg_label * Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression) list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val prop_get : loc:'a -> prop_loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> string -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val prop_set : loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> prop_loc:Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Ast_helper.loc -> Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> string -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val new_object : Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Longident.t Migrate_parsetree.OCaml_408.Ast.Asttypes.loc -> (Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Asttypes.arg_label * Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression) list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

Instantiation of a class, used by new%js.

module S : sig ... end

val format_meth : Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> Migrate_parsetree.OCaml_408.Ast.Parsetree.expression

We remove Pexp_poly as it should never be in the parsetree except after a method call.

module Prop_kind : sig ... end

type field_desc =

| Meth of string Migrate_parsetree.OCaml_408.Ast.Asttypes.loc * Migrate_parsetree.OCaml_408.Ast.Asttypes.private_flag * Migrate_parsetree.OCaml_408.Ast.Asttypes.override_flag * Migrate_parsetree.OCaml_408.Ast.Parsetree.expression * Arg.t list

| Val of string Migrate_parsetree.OCaml_408.Ast.Asttypes.loc * Prop_kind.t * Migrate_parsetree.OCaml_408.Ast.Asttypes.override_flag * Migrate_parsetree.OCaml_408.Ast.Parsetree.expression

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

val preprocess_literal_object : (Migrate_parsetree.OCaml_408.Ast.Parsetree.expression -> Migrate_parsetree.OCaml_408.Ast.Parsetree.expression) -> Migrate_parsetree.OCaml_408.Ast.Parsetree.class_field list -> [ `Error of Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.extension | `Fields of field_desc list ]

val literal_object : Migrate_parsetree.Ast_408.Parsetree.pattern -> field_desc list -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_408.Ast.Parsetree.expression

val mapper : Migrate_parsetree.OCaml_408.Ast.Ast_mapper.mapper